Format transactions into rows with action, amount and time.
e.g., Spent 0.00612687 2018-11-25 16:25

+++ app/src/main/java/com/blockchain/challenge/sources/BlockchainService.kt
@@ -10,7 +10,7 @@ interface BlockchainService {
     @GET("multiaddr")
     fun transactions(
             @Query("active") address : BitcoinAddress
-    ): Call<MultiAddressResponse> // Call -> Observable ?
+    ): Call<MultiAddressResponse>
 }
 
 data class MultiAddressResponse(
@@ -18,6 +18,6 @@ data class MultiAddressResponse(
 )
 
 data class BitcoinTransaction(
-    @SerializedName("result") val amount : Int,
-    val time : Int // TODO: decode to time
+    @SerializedName("result") val amount : Long,
+    val time : Long
 )
